add: one test for linux and windows
