+++ tests/end_to_end/test_vulnerability_detector/data/playbooks/configuration.yaml
@@ -2,7 +2,7 @@
   hosts: wazuh-agent
   tasks:
 
-    - name: Enable the agent module to collect installed packages
+    - name: Enable the agent module to collect installed packages (Ubuntu)
       become: true
       blockinfile:
         path: /var/ossec/etc/ossec.conf
@@ -16,9 +16,32 @@
           </wodle>
         marker: <!-- {mark} ANSIBLE MANAGED BLOCK -->
 
-    - name: Restart wazuh-agent
+    - name: Restart wazuh-agent (Ubuntu)
       become: true
       shell: systemctl restart wazuh-agent
+
+
+- name: Test agent configuration
+  hosts: wazuh-windows
+  tasks:
+
+    - name: Enable the agent module to collect installed packages (Windows)
+      ansible.windows.win_shell: |
+        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<ossec_config>"
+        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<wodle name='syscollector'>"
+        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<disabled>no</disabled>"
+        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<interval>10s</interval>"
+        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<os>yes</os>"
+        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<packages>yes</packages>"
+        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<hotfixes>yes</hotfixes>"
+        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n</wodle>"
+        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n</ossec_config>"
+
+    - name: Restart wazuh-agent (Windows)
+      ansible.windows.win_shell: |
+        net stop wazuh
+        net start wazuh
+
 
 - name: Test manager configuration
   hosts: wazuh-manager
@@ -47,30 +70,16 @@
           <update_interval>1h</update_interval>
           </provider>
 
-          <!-- Debian OS vulnerabilities -->
-          <provider name="debian">
-          <enabled>no</enabled>
-          <os>stretch</os>
-          <os>buster</os>
-          <update_interval>1h</update_interval>
-          </provider>
-
-          <!-- RedHat OS vulnerabilities -->
-          <provider name="redhat">
-          <enabled>no</enabled>
-          <os>8</os>
-          <update_interval>1h</update_interval>
-          </provider>
-
-          <!-- Windows OS vulnerabilities -->
+           <!-- Windows OS vulnerabilities -->
           <provider name="msu">
-          <enabled>no</enabled>
+          <enabled>yes</enabled>
           <update_interval>1h</update_interval>
           </provider>
+
           <!-- Aggregate vulnerabilities -->
           <provider name="nvd">
           <enabled>yes</enabled>
-          <update_from_year>2022</update_from_year>
+          <update_from_year>2021</update_from_year>
           <update_interval>1h</update_interval>
           </provider>
           </vulnerability-detector>
@@ -82,6 +91,7 @@
 
     - name: Wait until the feeds were downloaded and the first scan was completed
       become: true
-      ansible.builtin.wait_for:
+      wait_for:
         path: /var/ossec/logs/ossec.log
         search_regex: Vulnerability scan finished.
+
