+++ tests/end_to_end/test_vulnerability_detector/data/playbooks/generate_events.yaml
@@ -6,21 +6,28 @@
       shell: echo "" > /var/ossec/logs/alerts/alerts.json
       become: true
 
-- name: Generate events
+- name: Generate events linux
   hosts: wazuh-agent
   tasks:
 
-    - name: "{{ event_description }}"
+    - name: "{{ event_description_linux }}"
       become: true
-      command: "{{ command }}"
+      command: "{{ command_linux }}"
+
+- name: Generate events windows
+  hosts: wazuh-windows
+  tasks:
+
+    - name: "{{ event_description_windows }}"
+      ansible.windows.win_shell: "{{ command_windows }}"
+
+- name: Wait alert
+  hosts: wazuh-manager
+  tasks:
 
     - name: Wait for alert
       wait_for:
         timeout: 60
-
-- name: Get logs
-  hosts: wazuh-manager
-  tasks:
 
     - name: Get alerts.json
       fetch:
