+++ tests/end_to_end/test_vulnerability_detector/data/playbooks/teardown_playbook.yaml
@@ -1,0 +1,15 @@
+- name: Cleanup environment
+  hosts: wazuh-agent
+  tasks:
+
+    - name: Uninstall vim vulnerable package
+      become: true
+      command: apt-get remove vim -y
+
+- name: Cleanup environment
+  hosts: wazuh-windows
+  tasks:
+
+    - name: Uninstall Mozilla Firefox vulnerable package
+      ansible.windows.win_shell: |
+        Start-Process "C:\Program Files\Mozilla Firefox\uninstall\helper.exe" -ArgumentList "/S" -Wait
