+++ tests/end_to_end/test_vulnerability_detector/data/test_cases/cases_vulnerability_detector.yaml
@@ -1,0 +1,16 @@
+- name: detect_vulnerability
+  description: Detect vulnerability
+  configuration_parameters: null
+  metadata:
+    extra_vars:
+      event_description_linux: Install vulnerable vim package
+      command_linux: apt install -y vim=2:8.1.2269-1ubuntu5.7
+      event_description_windows: Install vulnerable Mozilla package
+      command_windows: |
+        Invoke-WebRequest -Uri "https://ftp.mozilla.org/pub/firefox/releases/92.0/win64/en-US/Firefox Setup 92.0.exe" `
+        -OutFile "C:\temp\test.exe"
+        Start-Process -Wait -FilePath "C:\temp\test.exe" -ArgumentList "/S" -PassThru
+    rule.id: 23505
+    rule.level: 10
+    rule.description_linux: CVE-2022-1621 affects vim
+    rule.description_windows: CVE-2021-30547 affects Mozilla Firefox \(x64 en-US\)
