--- tests/end_to_end/test_vulnerability_detector/test_vulnerability_detector_linux/data/playbooks/teardown_playbook.yaml
@@ -1,7 +1,0 @@
-- name: Cleanup environment
-  hosts: wazuh-agent
-  become: true
-  tasks:
-
-    - name: Uninstall vim vulnerable package
-      command: apt-get remove vim -y
