--- tests/end_to_end/test_vulnerability_detector/test_vulnerability_detector_linux/data/test_cases/cases_vulnerability_detector.yaml
@@ -1,10 +1,0 @@
-- name: vulnerability_detector
-  description: Detect vim vulnerability
-  configuration_parameters: null
-  metadata:
-    extra_vars:
-      event_description: Install vulnerable vim package
-      command: apt install vim=2:8.1.2269-1ubuntu5.7
-    rule.id: 23505
-    rule.level: 10
-    rule.description: CVE-2022-1621 affects vim
