--- tests/end_to_end/test_vulnerability_detector/test_vulnerability_detector_windows/data/playbooks/configuration.yaml
@@ -1,87 +1,0 @@
-- name: Test agent configuration
-  hosts: wazuh-windows
-  tasks:
-
-    - name: Enable the agent module to collect installed packages
-      ansible.windows.win_shell: |
-        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<ossec_config>"
-        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<wodle name='syscollector'>"
-        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<disabled>no</disabled>"
-        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<interval>10s</interval>"
-        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<os>yes</os>"
-        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<packages>yes</packages>"
-        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n<hotfixes>yes</hotfixes>"
-        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n</wodle>"
-        Add-Content 'C:\Program Files (x86)\ossec-agent\ossec.conf' "`n</ossec_config>"
-
-    - name: Restart wazuh-agent
-      ansible.windows.win_shell: |
-        net stop wazuh
-        net start wazuh
-
-- name: Test manager configuration
-  hosts: wazuh-manager
-  tasks:
-
-    - name: Truncate file ossec.log
-      shell: echo "" > /var/ossec/logs/ossec.log
-      become: true
-
-    - name: Enabled vulnerability detector module
-      become: true
-      blockinfile:
-        path: /var/ossec/etc/ossec.conf
-        insertbefore: </ossec_config>
-        block: |
-          <vulnerability-detector>
-          <enabled>yes</enabled>
-          <interval>10s</interval>
-          <min_full_scan_interval>10s</min_full_scan_interval>
-          <run_on_start>yes</run_on_start>
-
-          <!-- Ubuntu OS vulnerabilities -->
-          <provider name="canonical">
-          <enabled>no</enabled>
-          <os>focal</os>
-          <update_interval>1h</update_interval>
-          </provider>
-
-          <!-- Debian OS vulnerabilities -->
-          <provider name="debian">
-          <enabled>no</enabled>
-          <os>stretch</os>
-          <os>buster</os>
-          <update_interval>1h</update_interval>
-          </provider>
-
-          <!-- RedHat OS vulnerabilities -->
-          <provider name="redhat">
-          <enabled>no</enabled>
-          <os>8</os>
-          <update_interval>1h</update_interval>
-          </provider>
-
-          <!-- Windows OS vulnerabilities -->
-          <provider name="msu">
-          <enabled>yes</enabled>
-          <update_interval>1h</update_interval>
-          </provider>
-
-          <!-- Aggregate vulnerabilities -->
-          <provider name="nvd">
-          <enabled>yes</enabled>
-          <update_from_year>2021</update_from_year>
-          <update_interval>1h</update_interval>
-          </provider>
-          </vulnerability-detector>
-        marker: <!-- {mark} ANSIBLE MANAGED BLOCK -->
-
-    - name: Restart wazuh-manager
-      become: true
-      shell: systemctl restart wazuh-manager
-
-    - name: Wait until the feeds were downloaded and the first scan was completed
-      become: true
-      ansible.builtin.wait_for:
-        path: /var/ossec/logs/ossec.log
-        search_regex: Vulnerability scan finished.
