--- tests/end_to_end/test_vulnerability_detector/test_vulnerability_detector_windows/data/playbooks/generate_events.yaml
@@ -1,29 +1,0 @@
-- name: Truncate files
-  hosts: wazuh-manager
-  tasks:
-
-    - name: Truncate file alert.json
-      shell: echo "" > /var/ossec/logs/alerts/alerts.json
-      become: true
-
-- name: Generate events
-  hosts: wazuh-windows
-  tasks:
-
-    - name: "{{ event_description }}"
-      ansible.windows.win_shell: "{{ command }}"
-
-- name: Wait alert
-  hosts: wazuh-manager
-  tasks:
-
-    - name: Wait for alert
-      wait_for:
-        timeout: 60
-
-    - name: Get alerts.json
-      fetch:
-        src: /var/ossec/logs/alerts/alerts.json
-        dest: /tmp/
-        flat: true
-      become: true
