--- tests/end_to_end/test_vulnerability_detector/test_vulnerability_detector_windows/data/playbooks/teardown_playbook.yaml
@@ -1,7 +1,0 @@
-- name: Cleanup environment
-  hosts: wazuh-windows
-  tasks:
-
-    - name: Unistall Mozilla Firefox vulnerable package
-      ansible.windows.win_shell: |
-        Start-Process "C:\Program Files\Mozilla Firefox\uninstall\helper.exe" -ArgumentList "/S" -Wait
