--- tests/end_to_end/test_vulnerability_detector/test_vulnerability_detector_windows/data/test_cases/cases_vulnerability_detector.yaml
@@ -1,13 +1,0 @@
-- name: vulnerability_detector
-  description: Detect Mozilla vulnerability
-  configuration_parameters: null
-  metadata:
-    extra_vars:
-      event_description: Install vulnerable Mozilla package
-      command: |
-        Invoke-WebRequest -Uri "https://ftp.mozilla.org/pub/firefox/releases/92.0/win64/en-US/Firefox Setup 92.0.exe" `
-        -OutFile "C:\temp\test.exe"
-        Start-Process -Wait -FilePath "C:\temp\test.exe" -ArgumentList "/S" -PassThru
-    rule.id: 23505
-    rule.level: 10
-    rule.description: CVE-2021-30547 affects Mozilla Firefox \(x64 en-US\)
